+++ R/movAvg2.R
@@ -1,6 +1,6 @@
 ##' Calculate the moving average using a 2-sided, symmetric window
 ##'
-##' Wrapper for \code{smartFilter} that creates a set of symmetric weights for
+##' Wrapper for \code{\link{smartFilter}} that creates a set of symmetric weights for
 ##' the 2-sided window based on the Gaussian kernel, exponetial decay, linear
 ##' decay, or simple uniform weights, and then calculates the moving average
 ##' using those weights.