+++ R/stripExtension.R
@@ -1,7 +1,7 @@
 ##' Remove the extension of a vector of filenames
 ##' 
-##' Remove the extension of a vector of filenames, assuming that the extension is the set of characters that follows the last
-##' "."
+##' Remove the extension of a vector of filenames, assuming that the extension is the set of characters
+##' that follows the last \code{"."}
 ##'
 ##' Assumes paths are delineated using forward slashes.  If an \code{NA} is
 ##' supplied, then an \code{NA} is returned.  If the desired string doesn't