+++ tests/testthat/test_rma.R
@@ -7,6 +7,9 @@ test_that("rma() removes all objects from .GlobalEnv", {
    assign("y", c("this", "that"), envir = .GlobalEnv)
    assign("z", rnorm(20), envir = .GlobalEnv)
 
+   # Verify we have at least three objects in the global environment
+   expect_true(length(ls(envir = .GlobalEnv)) >= 3)
+   
    # Remove them all
    rma()